+++ duat-core/src/widgets/file.rs
@@ -117,7 +117,7 @@ pub struct File {
     path: Path,
     text: Text,
     cfg: PrintCfg,
-    printed_lines: Vec<(usize, bool)>,
+    printed_lines: Vec<(u32, bool)>,
 }
 
 impl File {
@@ -208,24 +208,24 @@ impl File {
     /// These are returned as a `usize`, showing the index of the line
     /// in the file, and a `bool`, which is `true` when the line is
     /// wrapped.
-    pub fn printed_lines(&self) -> &[(usize, bool)] {
+    pub fn printed_lines(&self) -> &[(u32, bool)] {
         &self.printed_lines
     }
 
     ////////// General querying functions
 
     /// The number of bytes in the file.
-    pub fn len_bytes(&self) -> usize {
+    pub fn len_bytes(&self) -> u32 {
         self.text.len().byte()
     }
 
     /// The number of [`char`]s in the file.
-    pub fn len_chars(&self) -> usize {
+    pub fn len_chars(&self) -> u32 {
         self.text.len().char()
     }
 
     /// The number of lines in the file.
-    pub fn len_lines(&self) -> usize {
+    pub fn len_lines(&self) -> u32 {
         self.text.len().line()
     }
